pyproject.toml
[project]
name = "context-engineering"
version = "0.1.0"
description = "Context Engineering Visualizer - Demonstrates information flow in LLM agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "gradio>=6.3.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
